introduces ProjectContext
As the final step of getting rid of MacroContext (the first step was
refactoring of error reporting capabilities for metaprograms), I've created
a dedicated abstraction for projects and moved resource acquisition functionality
there.
