+++ foundation/src/main/scala/org/scalameta/annotations/hosted.scala
@@ -4,7 +4,7 @@ import scala.language.experimental.macros
 import scala.annotation.StaticAnnotation
 import scala.reflect.macros.whitebox.Context
 
-class hosted(macroApi: Boolean = false) extends StaticAnnotation {
+class hosted extends StaticAnnotation {
   def macroTransform(annottees: Any*): Any = macro HostedMacros.impl
 }
 
@@ -15,12 +15,19 @@ class HostedMacros(val c: Context) {
       case q"new $_(..$args).macroTransform(..$_)" => args
       case q"new $_().macroTransform(..$_)" => Nil
     }
-    val macroApi = args.collect{ case q"macroApi = true" => true }.nonEmpty
-    val exnTpe = tq"_root_.scala.meta.MetaException"
-    val contextTpe = if (macroApi) tq"_root_.scala.meta.macros.Context" else tq"_root_.scala.meta.semantic.Context"
+    val contextTpt = {
+      def enclosingPackage(sym: Symbol): Symbol = {
+        if (sym == NoSymbol) sym
+        else if (sym.isPackage || sym.isPackageClass) sym
+        else enclosingPackage(sym.owner)
+      }
+      val apiSym = enclosingPackage(c.internal.enclosingOwner)
+      val apiRef = apiSym.fullName.split('.').foldLeft(q"_root_": Tree)((acc, part) => q"$acc.${TermName(part)}")
+      tq"$apiRef.Context"
+    }
     def transform(ddef: DefDef): DefDef = {
       val DefDef(mods, name, tparams, vparamss, tpt, body) = ddef
-      val contextful = q"new _root_.org.scalameta.annotations.contextful[$contextTpe]"
+      val contextful = q"new _root_.org.scalameta.annotations.contextful[$contextTpt]"
       val mods1 = Modifiers(mods.flags, mods.privateWithin, mods.annotations ++ List(contextful))
       DefDef(mods1, name, tparams, vparamss, tpt, body)
     }
