+++ scalameta/src/main/scala/scala/meta/macros/Api.scala
@@ -6,6 +6,4 @@ import org.scalameta.annotations._
 import scala.meta.macros.{Context => MacroContext}
 
 private[meta] trait Api {
-  @hosted(macroApi = true) def resources: Seq[String] = implicitly[MacroContext].resources.keys.toList.sortBy(Predef.identity _)
-  @hosted(macroApi = true) def resource(url: String): Array[Byte] = implicitly[MacroContext].resources(url)
 }
