+++ scalameta/src/main/scala/scala/meta/macros/Context.scala
@@ -7,5 +7,4 @@ import scala.annotation._
 @opaque
 @implicitNotFound("this method requires an implicit scala.meta.macros.Context")
 trait Context extends scala.meta.semantic.Context {
-  def resources: Map[String, Array[Byte]]
 }
