+++ scalameta/src/main/scala/scala/meta/projects/Project.scala
@@ -1,0 +1,11 @@
+package scala.meta
+package projects
+
+import scala.{Seq => _}
+import scala.collection.immutable.Seq
+
+trait Project {
+  def sources: Seq[Source]
+  def resources: Seq[Resource]
+  // TODO: later on, `def dependencies: Seq[Dependency]` etc
+}
