+++ scalameta/src/main/scala/scala/meta/projects/Resource.scala
@@ -1,0 +1,7 @@
+package scala.meta
+package projects
+
+trait Resource {
+  def url: String
+  def stream: java.io.InputStream
+}
