+++ scalameta/src/main/scala/scala/meta/syntactic/Content.scala
@@ -1,0 +1,10 @@
+package scala.meta
+package syntactic
+
+import scala.meta.internal.tokenizers._
+
+trait Content extends Input {
+  def chars: Array[Char]
+  private val tokenCache = scala.collection.mutable.Map[Dialect, Tokens]()
+  final override def tokens(implicit dialect: Dialect): Tokens = tokenCache.getOrElseUpdate(dialect, tokenize(this))
+}
